Extract spawn plumbing to lib/runtime/claude.js (verbatim move)
modelArgs, getBareArgs, getSpawnEnv, the child-pid registry (pid file +
recycling guard + register/unregister), resolveClaudeBin, killProcessTree,
and spawnClaude move verbatim. The registry moves WITH spawnClaude because
spawnClaude is its writer; the root's cleanup machinery (killAllChildren,
cleanOrphanedProcesses, the workspace-move clear) reads and prunes the same
file by calling in, root -> lib.

Count-asserted conversions: 6 WORKSPACE reads -> getWorkspace() at use time,
1 ACTUAL_PORT -> deps.getActualPort(). Everything else the module needs is
lib-owned (config, store persistence) or the repo-root codex.js probe
constant. Wire list is ONE entry: getActualPort (the port exists only after
server.listen()).

Pin retarget (Class B, enumerated): runtime-adapter.test.js 'Claude spawn
seam' now reads lib/runtime/claude.js. The legacy-spawn declaration-order
scan needs NO retarget: it anchors on the legacy chat branch call-site text,
which is WebSocket-handler territory and unmoved.

server.js 5,187 -> 4,958 lines; lib/runtime/claude.js 295 lines. Suite 1466.
